+++ src/main/java/com/joorani/myshop/controller/ProductController.java
@@ -1,7 +1,7 @@
 package com.joorani.myshop.controller;
 
-import com.joorani.myshop.entity.dtos.ProductRegisterForm;
-import com.joorani.myshop.entity.dtos.RegisteredProductDto;
+import com.joorani.myshop.controller.dtos.ProductRegisterForm;
+import com.joorani.myshop.controller.dtos.RegisteredProductDto;
 import com.joorani.myshop.service.ProductService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
