+++ src/main/java/com/joorani/myshop/controller/dtos/OptionDto.java
@@ -1,4 +1,4 @@
-package com.joorani.myshop.entity.dtos;
+package com.joorani.myshop.controller.dtos;
 
 import com.joorani.myshop.entity.ProductOption;
 import lombok.Builder;
