+++ src/main/java/com/joorani/myshop/service/ProductServiceImpl.java
@@ -3,9 +3,9 @@ package com.joorani.myshop.service;
 import com.joorani.myshop.common.exception.DBEmptyDataException;
 import com.joorani.myshop.entity.Product;
 import com.joorani.myshop.entity.Store;
-import com.joorani.myshop.entity.dtos.OptionDto;
-import com.joorani.myshop.entity.dtos.ProductRegisterForm;
-import com.joorani.myshop.entity.dtos.RegisteredProductDto;
+import com.joorani.myshop.controller.dtos.OptionDto;
+import com.joorani.myshop.controller.dtos.ProductRegisterForm;
+import com.joorani.myshop.controller.dtos.RegisteredProductDto;
 import com.joorani.myshop.repository.ProductRepository;
 import com.joorani.myshop.repository.StoreRepository;
 import lombok.RequiredArgsConstructor;
